Added files to use functions under the string.h header file
0-strcat explores the use of the function strcat which catenates two strings. 1-strncat uses a more advanced version of the strcat function strncat which catenates the 2nd string unto the second until a certain specified point. 2-strncpy shows the command strncpy which is used to copy one string to another until a certain point.
The function rather will continue to print null bytes if the limit isn't exhausted and the string being copied from is finished. Another function in file 3-strcmp.c called strcmp compares the two input strings. If they are the same the function returns 0, it returns a number less than 0 if the first string is less than the second and a number greater than 0 if the first string is greater than the second.
The file 4-rev_array contains a function capable of reversing the order of an array. 5-string_toupper.c converts a string to all uppercase characters, 6-cap_string.c capitalizes every word of a string while 7-leet encodes a string by changing some letters to numbers
The 7-leet does this using only two loops and 1 if statement. It loops the input string's characters, then loops over the search characters and checks if they are the same. If yes, it substitutes the character with the corresponding character in the encoding string.
'