Correcting error in 7-leet.c
Error in 7-leet.c due to use of strlen() function in comparison to an integer. This error arose from the fact that the output returned by strlen() is of type long unsigned int and therefore should not be compared to an integer.
It was solved by creating an integer variable and assigning to it the value of the strlen() function. This variable was then used in place of the function in the comparison.
Note: This error only arose with flag -Werror.