Done most tasks including advanced. Two left to be completed.
The first advanced task was to set the value of a pointer to a character. It was a function containing a simple assignment statement. The other problem resolved in this commit is the addition of the file 2-strchr which had not been previously added or committed. The remaining issues involve the modification of a program to make it portable to other situations and a file that contains the password of the crackme executable.
These would be the subjects of further commits.