My first hack
Successfully retrieved the password of a simple passworded file crackme2. The file was passworded or locked using strcmp, a function that compares two strings. I simply used 'ltrace ./crackme2' to get the name of the sought after environment variable, then used saved a test string into the environment variable. I then used the ltrace again to note the actions of strcmp and finally used an MD5 reverse dictionary to get the string equivalent of the sought after MD5 and used it. It was successful.