Successfully created my first two static libraries
The first one was done manually and I simply compiled the functions with the gcc option -c which converts files into object files with a .o ending. Then, I had to run the ar command with options -rc: r to update the files in the static library if it already exists and c to create it if it doesn't. The full command for the creation of the static library wasar -rc libmy.a *.o with '*.o' to specify that all object files in the directory are to be included in the library.

The second task was to write the commands for creation of a static library called liball.a in a bash script.