Fix return value error & add new files
Added file to add elements to the end of a list. This is done by creating the node in question, iterating to the last element in the list then changing the pointer there from NULL to the address of the node.
Also added file to free the heap memory allocated for the heap. This is done by using a pointer to iterate through all elements of the list and freeing them during the iteration. Then finally freeing the first pointer.
Finally, made updates to header file